Switch to Lists to represent parameters
It turned out it was a bit tricky to convert a Hashmap into a vector and
keep the order of the parameters correct. The new implementation is now
a bit more bulky and it is now up to the user to put the parameter's
values in the right order, i.e. the order that should be used for the
solver.